Paths are now iterable, created new DirectoryEnumerable protocol allowing me to drastically reduce boilerplate and duplicated code for recursive functions on Directory/OpenDirectory, PathCollection can now be instantiated from a DirectoryPath or OpenDirectory, moved more code into specific extensions/files
