+++ Sources/TrailBlazer/Open/OpenDirectory.swift
@@ -6,7 +6,13 @@ import Darwin
 
 public typealias OpenDirectory = Open<DirectoryPath>
 
-public extension Open where PathType: DirectoryPath {
+extension Open: DirectoryEnumerable where PathType: DirectoryPath {
+    public var dir: DIRType { return path.dir! }
+
+    func rewind() {
+        path.rewind()
+    }
+
     /**
     Retrieves the immediate children of the directory
 
@@ -17,328 +23,6 @@ public extension Open where PathType: DirectoryPath {
     public func children(options: DirectoryEnumerationOptions = []) -> PathCollection {
         // Since the directory is already opened, getting the immediate
         // children is always safe
-        return try! path.children(options: options)
-    }
-
-    /**
-    Recursively iterates through and retrives all children in all subdirectories
-
-    - Parameter depth: How many subdirectories may be recursively traversed (-1 for infinite depth)
-    - Parameter options: The options used while enumerating the children of the directory
-
-    - Returns: A PathCollection of all the files, directories, and other paths that are contained in self and its subdirectories
-    - Note: Opens any directories that are previously unopened and will close them afterwards if it was only opened for this API call
-
-    - Throws: `OpenDirectoryError.permissionDenied` when the calling process does not have access to the path
-    - Throws: `OpenDirectoryError.noProcessFileDescriptors` when the process has used all of its available file descriptors
-    - Throws: `OpenDirectoryError.noSystemFileDescriptors` when the entire system has run out of available file descriptors
-    - Throws: `OpenDirectoryError.pathDoesNotExist` when the path does not exist
-    - Throws: `OpenDirectoryError.outOfMemory` when there is not enough available memory to open the directory
-    - Throws: `OpenDirectoryError.pathNotDirectory` when the path you're trying to open exists and is not a directory. This should only occur if your DirectoryPath object was created before the path existed and then the path was created as a non-directory path type
-    */
-    public func recursiveChildren(depth: Int = -1, options: DirectoryEnumerationOptions = []) throws -> PathCollection {
-        return try path.recursiveChildren(depth: depth, options: options)
-    }
-
-    /**
-    Recursively changes the owner and group of all files and subdirectories
-
-    - Parameter owner: The uid of the owner of the path
-    - Parameter group: The gid of the group with permissions to access the path
-    - Parameter options: The options used while enumerating the children of the directory
-
-    - Throws: `ChangeOwnershipError.permissionDenied` when the calling process does not have the proper permissions to modify path ownership
-    - Throws: `ChangeOwnershipError.badAddress` when the path points to a location outside your addressible address space
-    - Throws: `ChangeOwnershipError.tooManySymlinks` when too many symlinks were encounter while resolving the path
-    - Throws: `ChangeOwnershipError.pathnameTooLong` when the path has more than `PATH_MAX` number of characters
-    - Throws: `ChangeOwnershipError.pathDoesNotExist` when the path does not exist
-    - Throws: `ChangeOwnershipError.noKernelMemory` when there is insufficient memory to change the path's ownership
-    - Throws: `ChangeOwnershipError.pathComponentNotDirectory` when a component of the path is not a directory
-    - Throws: `ChangeOwnershipError.readOnlyFileSystem` when the file system is in read-only mode
-    - Throws: `ChangeOwnershipError.ioError` when an I/O error occurred during the API call
-    - Throws: `OpenDirectoryError.permissionDenied` when the calling process does not have access to the path
-    - Throws: `OpenDirectoryError.noProcessFileDescriptors` when the process has used all of its available file descriptors
-    - Throws: `OpenDirectoryError.noSystemFileDescriptors` when the entire system has run out of available file descriptors
-    - Throws: `OpenDirectoryError.pathDoesNotExist` when the path does not exist
-    - Throws: `OpenDirectoryError.outOfMemory` when there is not enough available memory to open the directory
-    - Throws: `OpenDirectoryError.pathNotDirectory` when the path you're trying to open exists and is not a directory. This should only occur if your DirectoryPath object was created before the path existed and then the path was created as a non-directory path type
-    */
-    public func changeRecursive(owner uid: uid_t = ~0, group gid: gid_t = ~0, options: DirectoryEnumerationOptions = .includeHidden) throws {
-        let childPaths = children(options: options)
-
-        for file in childPaths.files {
-            try file.change(owner: uid, group: gid)
-        }
-
-        for path in childPaths.other {
-            try path.change(owner: uid, group: gid)
-        }
-
-        for directory in childPaths.directories {
-            try directory.changeRecursive(owner: uid, group: gid)
-        }
-
-        try change(owner: uid, group: gid)
-    }
-
-    /**
-    Recursively changes the owner and group of all files and subdirectories
-
-    - Parameter owner: The username of the owner of the path
-    - Parameter group: The name of the group with permissions to access the path
-    - Parameter options: The options used while enumerating the children of the directory
-
-    - Throws: `ChangeOwnershipError.permissionDenied` when the calling process does not have the proper permissions to modify path ownership
-    - Throws: `ChangeOwnershipError.badAddress` when the path points to a location outside your addressible address space
-    - Throws: `ChangeOwnershipError.tooManySymlinks` when too many symlinks were encounter while resolving the path
-    - Throws: `ChangeOwnershipError.pathnameTooLong` when the path has more than `PATH_MAX` number of characters
-    - Throws: `ChangeOwnershipError.pathDoesNotExist` when the path does not exist
-    - Throws: `ChangeOwnershipError.noKernelMemory` when there is insufficient memory to change the path's ownership
-    - Throws: `ChangeOwnershipError.pathComponentNotDirectory` when a component of the path is not a directory
-    - Throws: `ChangeOwnershipError.readOnlyFileSystem` when the file system is in read-only mode
-    - Throws: `ChangeOwnershipError.ioError` when an I/O error occurred during the API call
-    - Throws: `OpenDirectoryError.permissionDenied` when the calling process does not have access to the path
-    - Throws: `OpenDirectoryError.noProcessFileDescriptors` when the process has used all of its available file descriptors
-    - Throws: `OpenDirectoryError.noSystemFileDescriptors` when the entire system has run out of available file descriptors
-    - Throws: `OpenDirectoryError.pathDoesNotExist` when the path does not exist
-    - Throws: `OpenDirectoryError.outOfMemory` when there is not enough available memory to open the directory
-    - Throws: `OpenDirectoryError.pathNotDirectory` when the path you're trying to open exists and is not a directory. This should only occur if your DirectoryPath object was created before the path existed and then the path was created as a non-directory path type
-    */
-    public func changeRecursive(owner username: String? = nil, group groupname: String? = nil, options: DirectoryEnumerationOptions = .includeHidden) throws {
-        let uid: uid_t
-        let gid: gid_t
-
-        if let username = username {
-            uid = try getUserInfo(username).pw_uid
-        } else {
-            uid = ~0
-        }
-
-        if let groupname = groupname {
-            gid = try getGroupInfo(groupname).gr_gid
-        } else {
-            gid = ~0
-        }
-
-        try changeRecursive(owner: uid, group: gid, options: options)
-    }
-
-    /**
-    Recursively changes the permissions on all paths
-
-    - Parameter permissions: The new permissions for the paths
-    - Parameter options: The options used while enumerating the children of the directory
-
-    - Throws: `ChangePermissionsError.permissionDenied` when the calling process does not have the proper permissions to modify path permissions
-    - Throws: `ChangePermissionsError.badAddress` when the path points to a location outside your accessible address space
-    - Throws: `ChangePermissionsError.ioError` when an I/O error occurred during the API call
-    - Throws: `ChangePermissionsError.tooManySymlinks` when too many symlinks were encountered while resolving the path
-    - Throws: `ChangePermissionsError.pathnameTooLong` when the path has more than `PATH_MAX` number of characters
-    - Throws: `ChangePermissionsError.pathDoesNotExist` when the path does not exist
-    - Throws: `ChangePermissionsError.noKernelMemory` when there is insufficient memory to change the path's permissions
-    - Throws: `ChangePermissionsError.pathComponentNotDirectory` when a component of the path is not a directory
-    - Throws: `ChangePermissionsError.readOnlyFileSystem` when the file system is in read-only mode
-    - Throws: `OpenDirectoryError.permissionDenied` when the calling process does not have access to the path
-    - Throws: `OpenDirectoryError.noProcessFileDescriptors` when the process has used all of its available file descriptors
-    - Throws: `OpenDirectoryError.noSystemFileDescriptors` when the entire system has run out of available file descriptors
-    - Throws: `OpenDirectoryError.pathDoesNotExist` when the path does not exist
-    - Throws: `OpenDirectoryError.outOfMemory` when there is not enough available memory to open the directory
-    - Throws: `OpenDirectoryError.pathNotDirectory` when the path you're trying to open exists and is not a directory. This should only occur if your DirectoryPath object was created before the path existed and then the path was created as a non-directory path type
-    */
-    public func changeRecursive(permissions: FileMode, options: DirectoryEnumerationOptions = .includeHidden) throws {
-        let childPaths = children(options: options)
-
-        for file in childPaths.other {
-            try file.change(permissions: permissions)
-        }
-
-        for path in childPaths.other {
-            try path.change(permissions: permissions)
-        }
-
-        for directory in childPaths.directories {
-            try directory.changeRecursive(permissions: permissions)
-        }
-
-        try change(permissions: permissions)
-    }
-
-    /**
-    Recursively changes the permissions on all paths
-
-    - Parameters:
-        - owner: The permissions for the owner of the path
-        - group: The permissions for members of the group with access to the path
-        - others: The permissions for everyone else accessing the path
-        - bits: The gid, uid, and sticky bits of the path
-        - options: The options used while enumerating the children of the directory
-
-    - Throws: `ChangePermissionsError.permissionDenied` when the calling process does not have the proper permissions to modify path permissions
-    - Throws: `ChangePermissionsError.badAddress` when the path points to a location outside your accessible address space
-    - Throws: `ChangePermissionsError.ioError` when an I/O error occurred during the API call
-    - Throws: `ChangePermissionsError.tooManySymlinks` when too many symlinks were encountered while resolving the path
-    - Throws: `ChangePermissionsError.pathnameTooLong` when the path has more than `PATH_MAX` number of characters
-    - Throws: `ChangePermissionsError.pathDoesNotExist` when the path does not exist
-    - Throws: `ChangePermissionsError.noKernelMemory` when there is insufficient memory to change the path's permissions
-    - Throws: `ChangePermissionsError.pathComponentNotDirectory` when a component of the path is not a directory
-    - Throws: `ChangePermissionsError.readOnlyFileSystem` when the file system is in read-only mode
-    - Throws: `OpenDirectoryError.permissionDenied` when the calling process does not have access to the path
-    - Throws: `OpenDirectoryError.noProcessFileDescriptors` when the process has used all of its available file descriptors
-    - Throws: `OpenDirectoryError.noSystemFileDescriptors` when the entire system has run out of available file descriptors
-    - Throws: `OpenDirectoryError.pathDoesNotExist` when the path does not exist
-    - Throws: `OpenDirectoryError.outOfMemory` when there is not enough available memory to open the directory
-    - Throws: `OpenDirectoryError.pathNotDirectory` when the path you're trying to open exists and is not a directory. This should only occur if your DirectoryPath object was created before the path existed and then the path was created as a non-directory path type
-    */
-    public func changeRecursive(owner: FilePermissions, group: FilePermissions, others: FilePermissions, bits: FileBits, options: DirectoryEnumerationOptions = .includeHidden) throws {
-        try changeRecursive(permissions: FileMode(owner: owner, group: group, others: others, bits: bits), options: options)
-    }
-
-    /**
-    Recursively changes the permissions on all paths
-
-    - Parameters:
-        - owner: The permissions for the owner of the path
-        - group: The permissions for members of the group with access to the path
-        - others: The permissions for everyone else accessing the path
-        - bits: The gid, uid, and sticky bits of the path
-        - options: The options used while enumerating the children of the directory
-
-    - Throws: `ChangePermissionsError.permissionDenied` when the calling process does not have the proper permissions to modify path permissions
-    - Throws: `ChangePermissionsError.badAddress` when the path points to a location outside your accessible address space
-    - Throws: `ChangePermissionsError.ioError` when an I/O error occurred during the API call
-    - Throws: `ChangePermissionsError.tooManySymlinks` when too many symlinks were encountered while resolving the path
-    - Throws: `ChangePermissionsError.pathnameTooLong` when the path has more than `PATH_MAX` number of characters
-    - Throws: `ChangePermissionsError.pathDoesNotExist` when the path does not exist
-    - Throws: `ChangePermissionsError.noKernelMemory` when there is insufficient memory to change the path's permissions
-    - Throws: `ChangePermissionsError.pathComponentNotDirectory` when a component of the path is not a directory
-    - Throws: `ChangePermissionsError.readOnlyFileSystem` when the file system is in read-only mode
-    - Throws: `OpenDirectoryError.permissionDenied` when the calling process does not have access to the path
-    - Throws: `OpenDirectoryError.noProcessFileDescriptors` when the process has used all of its available file descriptors
-    - Throws: `OpenDirectoryError.noSystemFileDescriptors` when the entire system has run out of available file descriptors
-    - Throws: `OpenDirectoryError.pathDoesNotExist` when the path does not exist
-    - Throws: `OpenDirectoryError.outOfMemory` when there is not enough available memory to open the directory
-    - Throws: `OpenDirectoryError.pathNotDirectory` when the path you're trying to open exists and is not a directory. This should only occur if your DirectoryPath object was created before the path existed and then the path was created as a non-directory path type
-    */
-    public func changeRecursive(owner: FilePermissions? = nil, group: FilePermissions? = nil, others: FilePermissions? = nil, bits: FileBits? = nil, options: DirectoryEnumerationOptions = .includeHidden) throws {
-        let current = permissions
-        try changeRecursive(owner: owner ?? current.owner, group: group ?? current.group, others: others ?? current.others, bits: bits ?? current.bits, options: options)
-    }
-
-    /**
-    Recursively changes the permissions on all paths
-
-    - Parameters:
-        - ownerGroup: The permissions for the path owner and also members of the group with access to the path
-        - others: The permissions for everyone else accessing the path
-        - bits: The gid, uid, and sticky bits of the path
-        - options: The options used while enumerating the children of the directory
-
-    - Throws: `ChangePermissionsError.permissionDenied` when the calling process does not have the proper permissions to modify path permissions
-    - Throws: `ChangePermissionsError.badAddress` when the path points to a location outside your accessible address space
-    - Throws: `ChangePermissionsError.ioError` when an I/O error occurred during the API call
-    - Throws: `ChangePermissionsError.tooManySymlinks` when too many symlinks were encountered while resolving the path
-    - Throws: `ChangePermissionsError.pathnameTooLong` when the path has more than `PATH_MAX` number of characters
-    - Throws: `ChangePermissionsError.pathDoesNotExist` when the path does not exist
-    - Throws: `ChangePermissionsError.noKernelMemory` when there is insufficient memory to change the path's permissions
-    - Throws: `ChangePermissionsError.pathComponentNotDirectory` when a component of the path is not a directory
-    - Throws: `ChangePermissionsError.readOnlyFileSystem` when the file system is in read-only mode
-    - Throws: `OpenDirectoryError.permissionDenied` when the calling process does not have access to the path
-    - Throws: `OpenDirectoryError.noProcessFileDescriptors` when the process has used all of its available file descriptors
-    - Throws: `OpenDirectoryError.noSystemFileDescriptors` when the entire system has run out of available file descriptors
-    - Throws: `OpenDirectoryError.pathDoesNotExist` when the path does not exist
-    - Throws: `OpenDirectoryError.outOfMemory` when there is not enough available memory to open the directory
-    - Throws: `OpenDirectoryError.pathNotDirectory` when the path you're trying to open exists and is not a directory. This should only occur if your DirectoryPath object was created before the path existed and then the path was created as a non-directory path type
-    */
-    public func changeRecursive(ownerGroup perms: FilePermissions, others: FilePermissions? = nil, bits: FileBits? = nil, options: DirectoryEnumerationOptions = .includeHidden) throws {
-        let current = permissions
-        try changeRecursive(owner: perms, group: perms, others: current.others, bits: bits ?? current.bits, options: options)
-    }
-
-    /**
-    Recursively changes the permissions on all paths
-
-    - Parameters:
-        - ownerOthers: The permissions for the owner of the path and everyone else
-        - group: The permissions for members of the group with access to the path
-        - bits: The gid, uid, and sticky bits of the path
-        - options: The options used while enumerating the children of the directory
-
-    - Throws: `ChangePermissionsError.permissionDenied` when the calling process does not have the proper permissions to modify path permissions
-    - Throws: `ChangePermissionsError.badAddress` when the path points to a location outside your accessible address space
-    - Throws: `ChangePermissionsError.ioError` when an I/O error occurred during the API call
-    - Throws: `ChangePermissionsError.tooManySymlinks` when too many symlinks were encountered while resolving the path
-    - Throws: `ChangePermissionsError.pathnameTooLong` when the path has more than `PATH_MAX` number of characters
-    - Throws: `ChangePermissionsError.pathDoesNotExist` when the path does not exist
-    - Throws: `ChangePermissionsError.noKernelMemory` when there is insufficient memory to change the path's permissions
-    - Throws: `ChangePermissionsError.pathComponentNotDirectory` when a component of the path is not a directory
-    - Throws: `ChangePermissionsError.readOnlyFileSystem` when the file system is in read-only mode
-    - Throws: `OpenDirectoryError.permissionDenied` when the calling process does not have access to the path
-    - Throws: `OpenDirectoryError.noProcessFileDescriptors` when the process has used all of its available file descriptors
-    - Throws: `OpenDirectoryError.noSystemFileDescriptors` when the entire system has run out of available file descriptors
-    - Throws: `OpenDirectoryError.pathDoesNotExist` when the path does not exist
-    - Throws: `OpenDirectoryError.outOfMemory` when there is not enough available memory to open the directory
-    - Throws: `OpenDirectoryError.pathNotDirectory` when the path you're trying to open exists and is not a directory. This should only occur if your DirectoryPath object was created before the path existed and then the path was created as a non-directory path type
-    */
-    public func changeRecursive(ownerOthers perms: FilePermissions, group: FilePermissions? = nil, bits: FileBits? = nil, options: DirectoryEnumerationOptions = .includeHidden) throws {
-        let current = permissions
-        try changeRecursive(owner: perms, group: group ?? current.group, others: perms, bits: bits ?? current.bits, options: options)
-    }
-
-    /**
-    Recursively changes the permissions on all paths
-
-    - Parameters:
-        - groupOthers: The permissions for members of the group with access to the path and anyone else
-        - owner: The permissions for the owner of the path
-        - bits: The gid, uid, and sticky bits of the path
-        - options: The options used while enumerating the children of the directory
-
-    - Throws: `ChangePermissionsError.permissionDenied` when the calling process does not have the proper permissions to modify path permissions
-    - Throws: `ChangePermissionsError.badAddress` when the path points to a location outside your accessible address space
-    - Throws: `ChangePermissionsError.ioError` when an I/O error occurred during the API call
-    - Throws: `ChangePermissionsError.tooManySymlinks` when too many symlinks were encountered while resolving the path
-    - Throws: `ChangePermissionsError.pathnameTooLong` when the path has more than `PATH_MAX` number of characters
-    - Throws: `ChangePermissionsError.pathDoesNotExist` when the path does not exist
-    - Throws: `ChangePermissionsError.noKernelMemory` when there is insufficient memory to change the path's permissions
-    - Throws: `ChangePermissionsError.pathComponentNotDirectory` when a component of the path is not a directory
-    - Throws: `ChangePermissionsError.readOnlyFileSystem` when the file system is in read-only mode
-    - Throws: `OpenDirectoryError.permissionDenied` when the calling process does not have access to the path
-    - Throws: `OpenDirectoryError.noProcessFileDescriptors` when the process has used all of its available file descriptors
-    - Throws: `OpenDirectoryError.noSystemFileDescriptors` when the entire system has run out of available file descriptors
-    - Throws: `OpenDirectoryError.pathDoesNotExist` when the path does not exist
-    - Throws: `OpenDirectoryError.outOfMemory` when there is not enough available memory to open the directory
-    - Throws: `OpenDirectoryError.pathNotDirectory` when the path you're trying to open exists and is not a directory. This should only occur if your DirectoryPath object was created before the path existed and then the path was created as a non-directory path type
-    */
-    public func changeRecursive(groupOthers perms: FilePermissions, owner: FilePermissions? = nil, bits: FileBits? = nil, options: DirectoryEnumerationOptions = .includeHidden) throws {
-        let current = permissions
-        try changeRecursive(owner: owner ?? current.owner, group: perms, others: perms, bits: bits ?? current.bits, options: options)
-    }
-
-    /**
-    Recursively changes the permissions on all paths
-
-    - Parameters:
-        - ownerGroupOthers: The permissions for the owner of the path, members of the group, and everyone else
-        - bits: The gid, uid, and sticky bits of the path
-        - options: The options used while enumerating the children of the directory
-
-    - Throws: `ChangePermissionsError.permissionDenied` when the calling process does not have the proper permissions to modify path permissions
-    - Throws: `ChangePermissionsError.badAddress` when the path points to a location outside your accessible address space
-    - Throws: `ChangePermissionsError.ioError` when an I/O error occurred during the API call
-    - Throws: `ChangePermissionsError.tooManySymlinks` when too many symlinks were encountered while resolving the path
-    - Throws: `ChangePermissionsError.pathnameTooLong` when the path has more than `PATH_MAX` number of characters
-    - Throws: `ChangePermissionsError.pathDoesNotExist` when the path does not exist
-    - Throws: `ChangePermissionsError.noKernelMemory` when there is insufficient memory to change the path's permissions
-    - Throws: `ChangePermissionsError.pathComponentNotDirectory` when a component of the path is not a directory
-    - Throws: `ChangePermissionsError.readOnlyFileSystem` when the file system is in read-only mode
-    - Throws: `OpenDirectoryError.permissionDenied` when the calling process does not have access to the path
-    - Throws: `OpenDirectoryError.noProcessFileDescriptors` when the process has used all of its available file descriptors
-    - Throws: `OpenDirectoryError.noSystemFileDescriptors` when the entire system has run out of available file descriptors
-    - Throws: `OpenDirectoryError.pathDoesNotExist` when the path does not exist
-    - Throws: `OpenDirectoryError.outOfMemory` when there is not enough available memory to open the directory
-    - Throws: `OpenDirectoryError.pathNotDirectory` when the path you're trying to open exists and is not a directory. This should only occur if your DirectoryPath object was created before the path existed and then the path was created as a non-directory path type
-    */
-    public func changeRecursive(ownerGroupOthers perms: FilePermissions, bits: FileBits? = nil, options: DirectoryEnumerationOptions = .includeHidden) throws {
-        try changeRecursive(owner: perms, group: perms, others: perms, bits: bits ?? permissions.bits, options: options)
+        return PathCollection(self, options: options)
     }
 }
