+++ Tests/TrailBlazerTests/ChownTests.swift
@@ -26,7 +26,7 @@ class ChownTests: XCTestCase {
             XCTFail("Failed to get the home directory")
             return
         }
-        guard let file = FilePath(home + "\(UUID()).test") else {
+        guard var file = FilePath(home + "\(UUID()).test") else {
             XCTFail("Path is not a file")
             return
         }
@@ -59,7 +59,7 @@ class ChownTests: XCTestCase {
             XCTFail("Failed to get the home directory")
             return
         }
-        guard let file = FilePath(home + "\(UUID()).test") else {
+        guard var file = FilePath(home + "\(UUID()).test") else {
             XCTFail("Path is not a file")
             return
         }
@@ -93,7 +93,7 @@ class ChownTests: XCTestCase {
             XCTFail("Failed to get the home directory")
             return
         }
-        guard let file = FilePath(home + "\(UUID()).test") else {
+        guard var file = FilePath(home + "\(UUID()).test") else {
             XCTFail("Path is not a file")
             return
         }
@@ -121,7 +121,7 @@ class ChownTests: XCTestCase {
             XCTFail("Failed to get the home directory")
             return
         }
-        guard let file = FilePath(home + "\(UUID()).test") else {
+        guard var file = FilePath(home + "\(UUID()).test") else {
             XCTFail("Path is not a file")
             return
         }
